+++ lib/featurely.dart
@@ -13,5 +13,8 @@
 /// ```
 library;
 
+export 'src/api/api_exception.dart'
+    show FeaturelyApiException, FeaturelyErrorCode, FeaturelyNetworkException;
 export 'src/featurely_base.dart' show Featurely;
+export 'src/options.dart' show FeaturelyErrorListener;
 export 'src/theme.dart' show FeaturelyTheme;